feat: improve cmake build system and configuration
- Change fallback preference from shared to static libraries for simpler deployment
- Enhance multi-config generator support with Release mode default for optimal performance

+++ test/serialization/json/TESTS_JSONArrayEnumerator.cpp
@@ -200,8 +200,8 @@ namespace nfx::serialization::json::test
 		ArrayEnumerator enumerator( document );
 		ASSERT_TRUE( enumerator.setPath( "scores" ) );
 
-		std::vector<int> expectedScores = { 100, 95, 87, 92, 78 };
-		std::vector<int> actualScores;
+		std::vector<int64_t> expectedScores = { 100, 95, 87, 92, 78 };
+		std::vector<int64_t> actualScores;
 
 		// enumeratorate through all elements
 		while ( !enumerator.isEnd() )
